rosic.cpp: added includes from /others/ folder

+++ Libraries/JUCE/modules/rosic/others/rosic_KeyGenerator.cpp
@@ -65,7 +65,7 @@ void KeyGenerator::setLicenseeNameAndAddress(char *newLicenseeNameAndAddress)
   }
   if( newLicenseeNameAndAddress != NULL )
   {
-    int newLength = strlen(newLicenseeNameAndAddress);
+    int newLength = (int) strlen(newLicenseeNameAndAddress);
     licenseeNameAndAddress  = new char[newLength+1];
     for(int c=0; c<=newLength; c++) // the <= is valid here, because we have one more cell allocated
       licenseeNameAndAddress[c] = newLicenseeNameAndAddress[c];
